==> notebooks/1_get_dbGaP_abstracts.ipynb <==
# Get dbGaP Abstracts
This notebook retrieves the study titles and abstracts from the [dbGaP](https://www.ncbi.nlm.nih.gov/gap/) using a list of dbGaP accession numbers for the RADx-rad initiative.

**Author:** Peter W. Rose ([pwrose@ucsd.edu](mailto:pwrose@ucsd.edu))  
**Date:** 2025-03-13

import glob
import pandas as pd
import nih_utils

# Input files
DBGAP_IDS = "../data/studies.csv"
RADX_RAD_GRANTS = "../data/grants.csv"
# Output file
DBGAP_ABSTRACTS = "../derived_data/dbgap_abstracts.csv"

## Data Integration
Merge dbGaP data with grant data

dbgap = pd.read_csv(DBGAP_IDS, usecols=["project_num", "dbgap_accession"], dtype=str, keep_default_na=False)
grants = pd.read_csv(RADX_RAD_GRANTS, usecols=["project_num", "sub_project", "research_initiative"], dtype=str, keep_default_na=False)
dbgap = dbgap.merge(grants, on="project_num")

# Note, study phs002603.v1.p1 has two associated project numbers
dbgap["project_serial_num"] = dbgap["project_num"].apply(nih_utils.get_project_serial_num)
print(dbgap.shape[0])
dbgap

## Retrieve and Save dbGaP Titles and Abstracts

dbgap[['dbgap_accession', 'title', 'focus', 'description']] = dbgap['dbgap_accession'].apply(
    lambda accession: pd.Series(nih_utils.extract_dbgap_study_info(accession))
)
dbgap.head()

dbgap.to_csv(DBGAP_ABSTRACTS, index=False)
