==> notebooks/1_get_radx_mentions.ipynb <==
# 1_get_radx_mentions
This notebook queries full-text articles in PubMed Central for the keyword **"RADx"**. The retrieved publications include those related to the **RADx Initiative** as well as false positives, such as articles mentioning the [**RADX gene**](https://www.genenames.org/data/gene-symbol-report/#!/hgnc_id/HGNC:25486). 

The publications that mention RADx projects are used as a test set to validate the selection of RADx related publications (see 3_analyze_publications_likert.ipynb)

**Author:** Peter W. Rose ([pwrose@ucsd.edu](mailto:pwrose@ucsd.edu))  
**Date:** 2025-03-13

import requests
import json
import pandas as pd

RADX_RAD_MENTION = "../derived_data/publications_radx_mentions.csv"

project = "RADx"
start_date = "2020-01-01" # RADx projects started in 2020

# Query publications and preprints that mention the RADx project (source: Europe PMC)
url = f"https://www.ebi.ac.uk/europepmc/webservices/rest/search?query=%22{project}%22%20AND%20(FIRST_PDATE:{start_date})%20AND%20sort_date:y&resultType=core&pageSize=1000&format=json&cursorMark=*"

response = requests.get(url)
response.raise_for_status()
data = response.json()

df = pd.json_normalize(data["resultList"]["result"])
df = df.astype(str)
df.fillna("", inplace=True)
df.replace("nan", "", inplace=True)

df.rename(columns={"fullTextIdList.fullTextId": "fullTextId", "pmid": "pm_id", "pmcid": "pmc_id"}, inplace=True)

print(f"Number of publications mentioning {project}: {df.shape[0]}")

df.head()

df.to_csv(RADX_RAD_MENTION, index=False)
