==> notebooks/2_classify_publications_likert.ipynb <==
# RADx-rad Publication Relevance Classification Using Likert Scores and LLMs

This notebook classifies scientific publications for relevance to RADx-rad program objectives using Large Language Models (LLMs). 

Publications referencing RADx-rad grant numbers are evaluated to determine how closely they align with:

- **Funding Opportunity Announcements (FOAs)** objectives (based on full-text descriptions)
- **dbGaP Study Objectives** (based on study titles and abstracts)

LLMs assess relevance by comparing each publication's title and abstract against FOA descriptions and dbGaP study details, assigning a relevance score on a 5-point [Likert scale](https://www.scribbr.com/methodology/likert-scale/). Each LLM evaluation also includes a rationale explaining the assigned score.


## Workflow Overview

1. **Data Integration**
   - Load and merge preprocessed FOA, dbGaP, and publication datasets.

2. **Likert-Based Classification**
   - Evaluate publications against FOA and grant objectives using LLMs.
   - Annotate publications with Likert scores and detailed rationales.

**Author:** Peter W. Rose ([pwrose@ucsd.edu](mailto:pwrose@ucsd.edu))  
**Date:** 2025-03-13

import os
from pathlib import Path
import glob
import json
import time
import pandas as pd

import llm_utils

## Specify the LLM Model
Run this notebook multiple times to evaluate publications using different LLM models. 
Uncomment the desired model for each run.

# model = "meta-llama/Llama-3.3-70B-Instruct"
model = "deepseek-ai/DeepSeek-R1-Distill-Qwen-32B"
# model = "gpt-4o-mini"

FOA_DIR = "../derived_data/foas" # Path to full-text FOA documents
DBGAP_ABSTRACTS = "../derived_data/dbgap_abstracts.csv" # dbGaP data
RADX_RAD_ALL = "../derived_data/publications_pubmed_raw.csv" # List of publications that mention RADx-rad grant numbers
LAST_UPDATE = "2025-06-02" # Date publications were last retrieved from PubMed

LENGTH_THRESHOLD = 100 # minimum length of title + abstract
ANNOTATION = "annotation_full_text_5" # Name of subfolder to save Likert scoring annotations
RESULTS = "../results"

## 1. **Data Integration**
Load and merge the preprocessed FOA, dbGaP, and Publication data

### Load Funding Opportunity Announcement Documents for the RADx-rad Program

foas = pd.concat(
    (pd.read_csv(f).assign(sub_project=f.stem) for f in Path(FOA_DIR).glob("*.csv")),
    ignore_index=True
)
foas

### Load dbGaP Titles and Abstracts for the RADx-rad Program

dbgap_abstracts = pd.read_csv(DBGAP_ABSTRACTS, dtype=str, keep_default_na=False)
dbgap_abstracts.rename(columns={ "title": "dbgap_title", "description": "dbgap_abstract"}, inplace=True)
dbgap_abstracts.head(1)

### Load List of Publications that mention RADx-rad Grant Numbers

publications = pd.read_csv(RADX_RAD_ALL, usecols=["pm_id", "pmc_id", "doi", "title", "abstract", "keywords", "journal", "year", "authors", "article_type", "project_serial_num", "award_type", "supplement", "sub_project"], dtype=str, keep_default_na=False)

print(f"Number of publications: {publications.shape[0]}")

### Map the Publications to corresponding dbGaP Studies

publications = publications.merge(dbgap_abstracts, on=["project_serial_num", "sub_project"])
print(f"Number of publications: {publications.shape[0]}")
publications.drop_duplicates(inplace=True)
publications.head()

publications[publications["pmc_id"] == "PMC10784670"]

### Map the publications to the corresponding FOAs

publications = publications.merge(foas, on="sub_project")
print(f"Number of publications: {publications.shape[0]}")
publications.head()

publications[publications["pmc_id"] == "PMC10784670"]

## Run the LLM to evaluate Publications against FOA and Grant Objectives

prompt = '''
You are a researcher analyzing scientific publications. You will be provided with a FUNDING OPPORTUNITY description, a GRANT ABSTRACT, and a PUBLICATION ABSTRACT. 
Your task is to determine whether the publication is related to the research objectives of the funding opportunity and/or the objectives stated in the grant abstract. 
If the publication abstract is missing, use the title but only if it clearly matches the objectives.
Please provide your answer as a JSON object following these guidelines:

Rank your response based on the following Likert scale:
1: Strongly disagree
2: Disagree
3: Neither agree nor disagree
4: Agree
5: Strongly agree

- result: integer 1 to 5 on the Likert scale to indicate if the publication is related to the funding opportunity and/or the grant description.
- explanation: A clear, concise, and specific explanation of how you arrived at your conclusion. The explanation should describe why the publication is or is not relevant to the research objectives in the funding opportunity and grant's abstract.

Example JSON response:

```json
{
  "result": 5,
  "explanation": "The publication directly addresses the specific research aims and methodology outlined in the funding opportunity and grant's abstract."
}
'''

The response format directs the LLM to return the results in the proper JSON format.

response_format = {
    "type": "json_schema",
    "json_schema": {
        "name": "Likert-5",
        "schema": {
            "type": "object",
            "properties": {
                "result": {"type": "integer"},
                "explanation": {"type": "string"},
            },
            "required": ["result", "explanation"],
            "additionalProperties": False
        },
        "strict": True
    }
}

def eval_publication(row, model): 
    pm_id = row["pm_id"]
    dbgap_accession = row["dbgap_accession"]
    doi = row["doi"].replace("doi:", "").replace("/", "_")

    identifier = f"{pm_id}_{doi}_{dbgap_accession}"

    # Remove model prefixes such as meta-llama from the model string
    simple_model = model.split('/', 1)[-1]
    result_path = os.path.join(RESULTS, simple_model, ANNOTATION)
    os.makedirs(result_path, exist_ok=True)

    # Check if the result file already exists, then skip this record
    result_file = os.path.join(result_path, f"{identifier}.csv")
    if os.path.isfile(result_file):
        if (len(row["title"]) + len(row['abstract'])) < LENGTH_THRESHOLD:
            print(f"Remove {identifier}. Length of title + abstract < {LENGTH_THRESHOLD} characters")
            os.remove(result_file)
        return

    print(f"processing: {identifier}")

    if (len(row["title"]) + len(row['abstract'])) < LENGTH_THRESHOLD:
        print(f"Skipping {pm_id}. Length of title + abstract < {LENGTH_THRESHOLD} characters")
        return

    # If there is no abstract, use the title as the abstract to avoid LLM failures.
    publication_abstract = row['abstract']
    if publication_abstract == "":
        publication_abstract = row['title']

    # Create the context for the LLM
    context = f"FUNDING OPPORTUNITY:\n{row['summary']}\n\nGRANT ABSTRACT:\n{row['dbgap_title']}\n{row['dbgap_abstract']}\n\nPUBLICATION ABSTRACT:\n{row['title']}\n{publication_abstract}"
    start = time.time()
    #print(context)
    
    # Run the LLM
    response, usage = llm_utils.run_prompt(prompt, context, model, environment="development", response_format=response_format)
    end = time.time()

    # Parse the response
    data = json.loads(response)
    result = data.get("result", "1")
    explanation = data.get("explanation", "")

    # Get token usage and calculate cost  
    completion_tokens = usage["completion_tokens"]
    prompt_tokens = usage["prompt_tokens"]

    cost = llm_utils.get_token_cost(prompt_tokens, model, "input")
    cost += llm_utils.get_token_cost(completion_tokens, model, "output")

    # Collect result data
    elapsed_time = f"{end-start:.1f}"
    result_data = {
        'result': [result],
        'explanation': [explanation],
        'model': [simple_model],
        'prompt_tokens': [prompt_tokens],
        'completion_tokens': [completion_tokens],
        'elapsed_time': [elapsed_time],
        'cost': [cost]
    }

    # Append the result data to each row 
    pub_data = row.to_dict()
    merged_data = pub_data | result_data

    # Create and save the DataFrame
    df = pd.DataFrame(merged_data)
    df.to_csv(result_file, index=False)
    return

publications.apply(eval_publication, model=model, axis=1)
