==> notebooks/2_get_funding_opportuntities.ipynb <==
# Download the Full-Text Funding Opportunity Documents for the RADx-rad Program
This notebook downloads and parses the core Funding Opportunity Announcement (FOA) for the RADx-rad program.

**Author:** Peter W. Rose ([pwrose@ucsd.edu](mailto:pwrose@ucsd.edu))  
**Date:** 2025-03-13

import os
import requests
import pandas as pd
from bs4 import BeautifulSoup

# Input files
FOA_DIR = "../data/funding_opportunities.csv" # List of RADx-rad core funding opportunities
# Output file directory
DERIVED_DATA = "../derived_data"

## Load the list of FOAs

foa_info = pd.read_csv(FOA_DIR)
foa_info

## Download, parse, and save each FOA in a separate file

def load_document_from_url(url):
    response = requests.get(url)
    html_content = response.text

    # Parse the HTML document
    soup = BeautifulSoup(html_content, 'html.parser')
    # Extract all paragraphs
    paragraphs = [p.get_text() for p in soup.find_all('p')]
    return "/n".join(paragraphs)

def get_foa_full_text(row): 
    url = row["url"]
    identifier = row["id"]
    sub_project = row["sub_project"]

    print(f"parsing: {identifier}")

    result_path = os.path.join(DERIVED_DATA, "foas")
    os.makedirs(result_path, exist_ok=True)

    result_file = os.path.join(result_path, f"{sub_project}.csv")
    if os.path.isfile(result_file):
        return
        
    text = load_document_from_url(url)
    # Remove colons, they sometimes mislead the LLM
    text = text.replace(":", " ")

    result_data = {
        'summary': [text],
    }
    pub_data = row.to_dict()
    merged_data = pub_data | result_data

    # Create the DataFrame
    df = pd.DataFrame(merged_data)
    df.to_csv(result_file, index=False)

    print(f"Saving {sub_project} FOA to: {result_file}")
    return

_ = foa_info.apply(get_foa_full_text, axis=1)
